Moving sources from my personal repo into github.com/microsoft
There are no code changes in this. There are changes that need to be made,
but I wanted to move this to the correct place (github.com/microsoft) before
making those changes.

I wrote 100% of this code, and I wrote it while I was employed by Microsoft
in May 2020. Microsoft has all rights to this code.