Relax static lifetime bounds
The provider crate requires event and provider metadata to have
static lifetime, but I cannot see why this is necessary--ETW will
consume this metadata synchronously as part of the EventWrite call.

The macros crate will always ensure that this metadata is static,
which is a good practice when possible. But for dynamically-generated
events and providers, static is a problem.